data-engineering/requirements.txt
# Core
python-dotenv==1.0.0

# Database
pymysql==1.1.0
cryptography==41.0.7

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
Faker==20.1.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0

# Progress bar
tqdm>=4.66.0

# Date utilities
python-dateutil>=2.9.0